Change LongTimer to use `decimal` instead of `double` interval
instead of ~5 years max length, ~2.5 decillion years max length

+++ Irene/LongTimer.cs
@@ -5,8 +5,8 @@ namespace Irene;
 class LongTimer {
 	public bool AutoReset { get; set; }
 	public bool Enabled { get => _timer.Enabled; set => _timer.Enabled = value; }
-	public double Interval { get; private init; }
-	public double Remaining { get; private set; }
+	public decimal Interval { get; private init; }
+	public decimal Remaining { get; private set; }
 
 	public event ElapsedEventHandler? Elapsed;
 	protected virtual void OnElapsed(ElapsedEventArgs e) {
@@ -15,12 +15,12 @@ class LongTimer {
 	}
 
 	private readonly Timer _timer;
-	private double _period;
+	private decimal _period;
 
-	private const double _maxPeriod = int.MaxValue - 1;
-	private const double _accuracy = 20; // msec
+	private const decimal _maxPeriod = int.MaxValue - 1;
+	private const decimal _accuracy = 20; // msec
 
-	public LongTimer(double totalMilliseconds, bool autoReset=false) {
+	public LongTimer(decimal totalMilliseconds, bool autoReset=false) {
 		Interval = totalMilliseconds;
 		Remaining = Interval;
 		Elapsed = null;
@@ -43,13 +43,14 @@ class LongTimer {
 
 			// If continuing, set next timer iteration and resume.
 			_period = NextPeriod();
-			_timer.Interval = _period;
+			_timer.Interval = (double)_period;
 			_timer.Start();
 		}
 
 		// Set up sub-timer.
 		_period = NextPeriod();
-		_timer = new Timer(_period) { AutoReset = autoReset };
+		_timer = new Timer((double)_period)
+			{ AutoReset = autoReset };
 		_timer.Elapsed += SetNextTimer;
 	}
 
@@ -66,7 +67,7 @@ class LongTimer {
 	public void Stop() => _timer.Stop();
 	public void Close() => _timer.Close();
 
-	private double NextPeriod() =>
+	private decimal NextPeriod() =>
 		(Remaining > _maxPeriod)
 			? Interval
 			: _maxPeriod;
